+++ src/RcppExports.cpp
@@ -620,12 +620,13 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// initialize_interrupt_handler
-void initialize_interrupt_handler();
-RcppExport SEXP _reticulate_initialize_interrupt_handler() {
+// py_interrupt_handler
+void py_interrupt_handler(int signum);
+RcppExport SEXP _reticulate_py_interrupt_handler(SEXP signumSEXP) {
 BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
-    initialize_interrupt_handler();
+    Rcpp::traits::input_parameter< int >::type signum(signumSEXP);
+    py_interrupt_handler(signum);
     return R_NilValue;
 END_RCPP
 }
@@ -685,7 +686,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_reticulate_r_convert_date", (DL_FUNC) &_reticulate_r_convert_date, 2},
     {"_reticulate_py_set_interrupt_impl", (DL_FUNC) &_reticulate_py_set_interrupt_impl, 0},
     {"_reticulate_readline", (DL_FUNC) &_reticulate_readline, 1},
-    {"_reticulate_initialize_interrupt_handler", (DL_FUNC) &_reticulate_initialize_interrupt_handler, 0},
+    {"_reticulate_py_interrupt_handler", (DL_FUNC) &_reticulate_py_interrupt_handler, 1},
     {NULL, NULL, 0}
 };